langchain_structured_output/requirements.txt
langchain
langchain-groq
python-dotenv
requests
pydantic
pydantic[email]
